+++ src/onediff/infer_compiler/__init__.py
@@ -2,7 +2,7 @@ import os
 import torch
 import oneflow as flow
 
-from .patch_for_compiling import *  # TODO:
+from .utils.patch_for_compiler import *  # TODO:
 from .registry import register
 from .with_oneflow_compile import oneflow_compile
