+++ src/onediff/infer_compiler/utils/patch_for_compiler.py
